Applies custom theme to the application
Replaces the default dark theme with a custom theme, "abrelTheme," to align the application's appearance with the brand's design guidelines.

This includes adjustments to primary, success, warning, error and info colors as well as changes to button, input, layout, table, tooltip and message components for a more cohesive user experience.